refactor(app.js): Update import paths for server and swagger files
