+++ backend/api/controllers/equipamentosController.js
