+++ backend/api/routes/equipamentos.js
