--- backend/src/db/comentarios.json
@@ -1,72 +1,0 @@
-[
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "validado": true,
-    "resposta": "Faremos o possível para resolver isso",
-    "id": "042se6OXW7C"
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido,teste 2",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "validado": true,
-    "resposta": "Faremos o possível para resolver isso",
-    "id": "04Ooc-ri2Db"
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido,teste 3,sem status da validação explicita",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "resposta": "",
-    "id": "04WPxbsqdI3",
-    "validado": true
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido,teste 3,sem status da validação explicita",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "resposta": "Faremos o possível para resolver isso",
-    "id": "04Rb9AR2NA7",
-    "validado": true
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido,teste 3,sem status da validação explicita",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "resposta": "Faremos o possível para resolver isso",
-    "id": "04bWXX-dOhb",
-    "validado": true
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sobre a sala,com tudo preenchido,teste 3,sem status da validação explicita",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "resposta": "Isso vai ser resolvido rapidamente",
-    "validado": true,
-    "id": "042KSyQC1hc"
-  },
-  {
-    "id_post": "123",
-    "id_usuario": "456",
-    "texto": "Comentário sem sentido",
-    "data": "2023-06-15T12:00:00Z",
-    "destinatario": "administrador",
-    "resposta": "",
-    "validado": false,
-    "id": "04txUPnKlwm"
-  }
-]
